+++ packaging/transports/__init__.py
@@ -1,0 +1,12 @@
+# This file is dual licensed under the terms of the Apache License, Version
+# 2.0, and the BSD License. See the LICENSE file in the root of this repository
+# for complete details.
+from __future__ import absolute_import, division, print_function
+
+
+from .wrappers import Transports  # noqa
+
+try:
+    from .requests import RequestsTransport  # noqa
+except ImportError:
+    pass
